# mcp/src/flagsmith_mcp/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flagsmith API",
    "version": "v1",
    "description": "Flagsmith's Core and SDK APIs. Check out <a href='https://docs.flagsmith.com'>Flagsmith documentation</a>.",
    "contact": {
      "email": "support@flagsmith.com"
    },
    "license": {
      "name": "BSD License",
      "identifier": "BSD-3-Clause"
    }
  },
  "servers": [
    {
      "url": "https://api.flagsmith.com",
      "description": "Flagsmith API"
    }
  ],
  "paths": {
...
    "/api/v1/projects/{id}/": {
      "get": {
        "operationId": "get_project",
        "description": "Retrieves comprehensive information about a specific project including configuration and statistics.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A unique integer value identifying this project.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRetrieve"
                }
              }
            }
          }
        },
        "security": [
          {
            "tokenAuth": []
          },
          {
            "Master API Key": []
          }
        ],
        "tags": [
          "mcp"
        ]
      },
      "put": {
        "operationId": "update_project",
        "description": "Updates project configuration settings such as the project name and feature visibility.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A unique integer value identifying this project.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectUpdate"
                }
              }
            }
          }
        },
        "security": [
          {
            "tokenAuth": []
          },
          {
            "Master API Key": []
          }
        ],
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/v1/projects/{id}/environments/": {
      "get": {
        "operationId": "list_project_environments",
        "description": "Retrieves all environments configured for the specified project.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A unique integer value identifying this project.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectList"
                }
              }
            }
          }
        },
        "security": [
          {
            "tokenAuth": []
          },
          {
            "Master API Key": []
          }
        ],
        "tags": [
          "mcp"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EdgeV2MigrationStatusEnum": {
        "description": "* `NOT_STARTED` - Not Started\n* `IN_PROGRESS` - In Progress\n* `COMPLETE` - Complete\n* `INCOMPLETE` - Incomplete (identity overrides skipped)",
        "type": "string",
        "enum": [
          "NOT_STARTED",
          "IN_PROGRESS",
          "COMPLETE",
          "INCOMPLETE"
        ]
      },
...
      "ProjectList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "uuid": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "maxLength": 2000
          },
          "organisation": {
            "type": "integer"
          },
          "hide_disabled_flags": {
            "description": "If true will exclude flags from SDK which are disabled",
            "type": "boolean"
          },
          "enable_dynamo_db": {
            "description": "If true will sync environment data with dynamodb and allow access to dynamodb identities",
            "type": "boolean",
            "readOnly": true
          },
          "migration_status": {
            "description": "Edge migration status of the project; can be one of: MIGRATION_SCHEDULED, MIGRATION_COMPLETED, MIGRATION_IN_PROGRESS, MIGRATION_NOT_STARTED, NOT_APPLICABLE",
            "type": "string",
            "readOnly": true
          },
          "use_edge_identities": {
            "type": "boolean",
            "readOnly": true
          },
          "prevent_flag_defaults": {
            "description": "Prevent defaults from being set in all environments when creating a feature.",
            "type": "boolean"
          },
          "enable_realtime_updates": {
            "description": "Enable this to trigger a realtime(sse) event whenever the value of a flag changes",
            "type": "boolean"
          },
          "only_allow_lower_case_feature_names": {
            "description": "Used by UI to validate feature names",
            "type": "boolean"
          },
          "feature_name_regex": {
            "description": "Used for validating feature names",
            "type": [
              "string",
              "null"
            ],
            "maxLength": 255
          },
          "show_edge_identity_overrides_for_feature": {
            "type": "boolean",
            "readOnly": true
          },
          "stale_flags_limit_days": {
            "description": "Number of days without modification in any environment before a flag is considered stale.",
            "type": "integer",
            "maximum": 2147483647,
            "minimum": -2147483648
          },
          "edge_v2_migration_status": {
            "description": "[Edge V2 migration] Project migration status. Set to `IN_PROGRESS` to trigger migration start.\n\n* `NOT_STARTED` - Not Started\n* `IN_PROGRESS` - In Progress\n* `COMPLETE` - Complete\n* `INCOMPLETE` - Incomplete (identity overrides skipped)",
            "allOf": [
              {
                "$ref": "#/components/schemas/EdgeV2MigrationStatusEnum"
              }
            ],
            "readOnly": true
          },
          "minimum_change_request_approvals": {
            "type": [
              "integer",
              "null"
            ],
            "maximum": 2147483647,
            "minimum": -2147483648
          },
          "enforce_feature_owners": {
            "description": "Require at least one user or group owner when creating a feature.",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "organisation"
        ]
      },
      "ProjectRetrieve": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "uuid": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "maxLength": 2000
          },
          "organisation": {
            "type": "integer"
          },
          "hide_disabled_flags": {
            "description": "If true will exclude flags from SDK which are disabled",
            "type": "boolean"
          },
          "enable_dynamo_db": {
            "description": "If true will sync environment data with dynamodb and allow access to dynamodb identities",
            "type": "boolean"
          },
          "migration_status": {
            "description": "Edge migration status of the project; can be one of: MIGRATION_SCHEDULED, MIGRATION_COMPLETED, MIGRATION_IN_PROGRESS, MIGRATION_NOT_STARTED, NOT_APPLICABLE",
            "type": "string",
            "readOnly": true
          },
          "use_edge_identities": {
            "type": "boolean",
            "readOnly": true
          },
          "prevent_flag_defaults": {
            "description": "Prevent defaults from being set in all environments when creating a feature.",
            "type": "boolean"
          },
          "enable_realtime_updates": {
            "description": "Enable this to trigger a realtime(sse) event whenever the value of a flag changes",
            "type": "boolean"
          },
          "only_allow_lower_case_feature_names": {
            "description": "Used by UI to validate feature names",
            "type": "boolean"
          },
          "feature_name_regex": {
            "description": "Used for validating feature names",
            "type": [
              "string",
              "null"
            ],
            "maxLength": 255
          },
          "show_edge_identity_overrides_for_feature": {
            "type": "boolean",
            "readOnly": true
          },
          "stale_flags_limit_days": {
            "description": "Number of days without modification in any environment before a flag is considered stale.",
            "type": "integer",
            "maximum": 2147483647,
            "minimum": -2147483648
          },
          "edge_v2_migration_status": {
            "description": "[Edge V2 migration] Project migration status. Set to `IN_PROGRESS` to trigger migration start.\n\n* `NOT_STARTED` - Not Started\n* `IN_PROGRESS` - In Progress\n* `COMPLETE` - Complete\n* `INCOMPLETE` - Incomplete (identity overrides skipped)",
            "allOf": [
              {
                "$ref": "#/components/schemas/EdgeV2MigrationStatusEnum"
              }
            ]
          },
          "minimum_change_request_approvals": {
            "type": [
              "integer",
              "null"
            ],
            "maximum": 2147483647,
            "minimum": -2147483648
          },
          "enforce_feature_owners": {
            "description": "Require at least one user or group owner when creating a feature.",
            "type": "boolean"
          },
          "max_segments_allowed": {
            "description": "Max segments allowed for this project",
            "type": "integer",
            "readOnly": true
          },
          "max_features_allowed": {
            "description": "Max features allowed for this project",
            "type": "integer",
            "readOnly": true
          },
          "max_segment_overrides_allowed": {
            "description": "Max segments overrides allowed for any (one) environment within this project",
            "type": "integer",
            "readOnly": true
          },
          "total_features": {
            "type": "integer",
            "readOnly": true
          },
          "total_segments": {
            "type": "integer",
            "readOnly": true
          }
        },
        "required": [
          "name",
          "organisation"
        ]
      },
      "ProjectUpdate": {
        "description": "Mixin to add read only status to fields in a given serializer based on the existence\nof a subscription and a black list of plan ids\n\nExample usage:\n\n    class MySerializer(ReadOnlyIfNotValidPlanMixin, ModelSerializer):\n        class Meta:\n            model = MyModel\n            fields = (\"my_field\",)\n\n        invalid_plans = (\"free\",)\n        field_names = (\"my_field\",)\n\n        def get_subscription(self):\n            return subscription",
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "readOnly": true
          },
          "uuid": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "maxLength": 2000
          },
          "organisation": {
            "type": "integer",
            "readOnly": true
          },
          "hide_disabled_flags": {
            "description": "If true will exclude flags from SDK which are disabled",
            "type": "boolean"
          },
          "enable_dynamo_db": {
            "description": "If true will sync environment data with dynamodb and allow access to dynamodb identities",
            "type": "boolean",
            "readOnly": true
          },
          "migration_status": {
            "description": "Edge migration status of the project; can be one of: MIGRATION_SCHEDULED, MIGRATION_COMPLETED, MIGRATION_IN_PROGRESS, MIGRATION_NOT_STARTED, NOT_APPLICABLE",
            "type": "string",
            "readOnly": true
          },
          "use_edge_identities": {
            "type": "boolean",
            "readOnly": true
          },
          "prevent_flag_defaults": {
            "description": "Prevent defaults from being set in all environments when creating a feature.",
            "type": "boolean"
          },
          "enable_realtime_updates": {
            "description": "Enable this to trigger a realtime(sse) event whenever the value of a flag changes",
            "type": "boolean",
            "readOnly": true
          },
          "only_allow_lower_case_feature_names": {
            "description": "Used by UI to validate feature names",
            "type": "boolean"
          },
          "feature_name_regex": {
            "description": "Used for validating feature names",
            "type": [
              "string",
              "null"
            ],
            "maxLength": 255
          },
          "show_edge_identity_overrides_for_feature": {
            "type": "boolean",
            "readOnly": true
          },
          "stale_flags_limit_days": {
            "description": "Number of days without modification in any environment before a flag is considered stale.",
            "type": "integer",
            "maximum": 2147483647,
            "minimum": -2147483648,
            "readOnly": true
          },
          "edge_v2_migration_status": {
            "description": "[Edge V2 migration] Project migration status. Set to `IN_PROGRESS` to trigger migration start.\n\n* `NOT_STARTED` - Not Started\n* `IN_PROGRESS` - In Progress\n* `COMPLETE` - Complete\n* `INCOMPLETE` - Incomplete (identity overrides skipped)",
            "allOf": [
              {
                "$ref": "#/components/schemas/EdgeV2MigrationStatusEnum"
              }
            ],
            "readOnly": true
          },
          "minimum_change_request_approvals": {
            "type": [
              "integer",
              "null"
            ],
            "maximum": 2147483647,
            "minimum": -2147483648
          },
          "enforce_feature_owners": {
            "description": "Require at least one user or group owner when creating a feature.",
            "type": "boolean"
          }
        },
        "required": [
          "name"
        ]
      },
...
    },
    "securitySchemes": {
      "Cohort Sync Key": {
        "type": "http",
        "scheme": "bearer",
        "description": "For cohort sync endpoints called by an external cohort source, such as Amplitude."
      },
      "Cohort Sync Key (Basic)": {
        "type": "http",
        "scheme": "basic",
        "description": "For cohort sync endpoints called by an external cohort source that can only send Basic credentials, such as Mixpanel. The key is the password; the username is ignored."
      },
      "Environment API Key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Environment-Key",
        "description": "For SDK endpoints. <a href='https://docs.flagsmith.com/clients/rest#public-api-endpoints'>Find out more</a>."
      },
      "Master API Key": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "For Management API endpoints. <a href='https://docs.flagsmith.com/integrating-with-flagsmith/flagsmith-api-overview/management-api'>Find out more</a>."
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "tokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Token-based authentication with required prefix \"Token\""
      }
    }
  },
...
}